fix(worker): stop hook imports from entering codeExec worker bundle
Extract getMatrixChildArrayArgs into entities/argument/lib so
createRuntimeMatrix no longer imports useArgumentsParsing (which pulls
tRPC/Next.js via useSolutionParameterNames). Import stringifySolutionResult
from its module path in the worker.
