+++ src/entities/argument/lib/index.ts
@@ -4,3 +4,7 @@ export { isArgumentArrayType } from "./isArgumentArrayType";
 export { isArgumentTreeType } from "./isArgumentTreeType";
 export { isArgumentObjectValid } from "./isArgumentObjectValid";
 export { getArgumentDisplayLabel } from "./getArgumentDisplayLabel";
+export {
+  getChildArrayName,
+  getMatrixChildArrayArgs,
+} from "./getMatrixChildArrayArgs";
